+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [connect] Add method `add_to_queue` to `Spirc` to add tracks, episodes, albums and playlists to the queue
+- [playback] Add `AddedToQueue` player event, emitting when a track was added to the queue with `Spirc::add_to_queue`
+
+### Changed
+
+- [core] Made `SpotifyId::to_base62`, `SpotifyId::to_base16`, `FileId::to_base16`, `SpotifyUri::to_id`, `SpotifyUri::to_uri` infallible (breaking)
+
+### Fixed
+
+- [audio] Fixed integer overflow in throughput calculation
+- [main] Fixed `--volume-ctrl fixed` not disabling volume control
+- [core] Fix default permissions on credentials file and warn user if file is world readable
+- [core] Try all resolved addresses for the dealer connection instead of failing after the first one.
+
 ## [0.8.0] - 2025-11-10
 
 ### Added